+++ internal/cli/bumpers.yml
@@ -1,0 +1,25 @@
+rules:
+    - match:
+        pattern: ^go test
+      send: |-
+        Use "just test" instead for TDD integration:
+        - just test                        # Run all tests
+        - just test PKG=./internal/claude  # Test only a specific package
+        See justfile for more information.
+    - match:
+        pattern: ^(gci|go vet|goimports|gofumpt|go fmt)
+      send: Use "just lint fix" instead to resolve lint/formatting issues.
+    - match:
+        pattern: ^cd /tmp
+      send: Create a "tmp" directory in the project root instead.
+commands:
+    - name: ""
+      send: Available commands:\n!help - Show this help\n!status - Show project status\n!docs - Open documentation
+    - name: ""
+      send: 'Project Status: All systems operational'
+    - name: ""
+      send: "\U0001F4DA Documentation: Visit https://github.com/wizzomafizzo/bumpers for usage guides and examples"
+session:
+    - add: Use 'just test' instead of 'go test' for proper TDD integration
+    - add: Check CLAUDE.md for project conventions and guidelines
+    - add: Run 'just lint fix' to resolve formatting and linting issues
