+++ internal/platform/claude/launcher_integration_test.go
@@ -3,7 +3,6 @@
 package claude
 
 import (
-	"context"
 	"testing"
 
 	"github.com/wizzomafizzo/bumpers/internal/testing"
